Restructure a bit and sort out tagline
