+++ pages/index.vue
@@ -3,7 +3,7 @@
 </template>
 <script>
 import BusinessPage from '@/components/BusinessPage'
-import { REGION_LONDON } from '@/utils'
+import { REGION_LONDON } from '@/regions'
 
 export default {
   components: { BusinessPage },
@@ -36,10 +36,7 @@ export default {
     }
   },
   head() {
-    return this.buildHead(
-      'Repair Directory',
-      'Find a London business to repair your broken devices.'
-    )
+    return this.buildHead('Repair Directory', this.tagline)
   },
 }
 </script>
