+++ regions.js
@@ -1,6 +1,12 @@
 export const REGION_LONDON = 'London'
 export const REGION_WALES = 'Wales'
 
+export const TAGLINE_GENERIC = 'Find a business to repair your broken devices.'
+export const TAGLINE_LONDON =
+  'Find a London business to repair your broken devices.'
+export const TAGLINE_WALES =
+  'Find a Welsh business to repair your broken devices.'
+
 // The bounds control which businesses we want to show.  At the moment we only support London and Wales, which
 // are sufficiently far apart that we can distinguish between them using a box rather than a polygon.
 export const BOUNDS_LONDON = {
